+++ apps/admin/app/students/_components/StudentList.tsx
@@ -1,15 +1,9 @@
-import { AwardIcon, StarCheckIcon, Text } from "@wow-class/ui";
-import Link from "next/link";
-import type { CSSProperties } from "react";
-import type {
-  StudyStudentApiResponseDto,
-  StudyTaskResponseDto,
-} from "types/dtos/studyStudent";
-import { formatNumberToPercent } from "utils/formatNumber";
+import { Text } from "@wow-class/ui";
+import type { StudyStudentApiResponseDto } from "types/dtos/studyStudent";
 import Table from "wowds-ui/Table";
-import TextButton from "wowds-ui/TextButton";
 
-import TaskTag from "./TaskTag";
+import StudentListItem from "./StudentListItem";
+import { StudyTasksThs } from "./StudyTasks";
 
 const STUENT_INFO_LIST_BEFORE = [
   "수료",
@@ -45,107 +39,13 @@ const StudentList = ({
       </Table.Thead>
       <Table.Tbody>
         {studentList.map((student) => (
-          <StudentListItem key={student.memberId} {...student} />
+          <Table.Tr key={student.memberId} value={student.memberId}>
+            <StudentListItem {...student} />
+          </Table.Tr>
         ))}
       </Table.Tbody>
     </Table>
   );
 };
 
-const StudentListItem = ({
-  memberId,
-  studyHistoryStatus,
-  isFirstRoundOutstandingStudent,
-  isSecondRoundOutstandingStudent,
-  name,
-  studentId,
-  discordUsername,
-  nickname,
-  githubLink,
-  studyTasks,
-  assignmentRate,
-  attendanceRate,
-}: StudyStudentApiResponseDto) => {
-  return (
-    <Table.Tr key={memberId} value={memberId}>
-      <Table.Td>
-        <StarCheckIcon checked={studyHistoryStatus === "COMPLETED"} />
-      </Table.Td>
-      <Table.Td>
-        <Text style={awardTextStyle} typo="body2">
-          <AwardIcon disabled={!isFirstRoundOutstandingStudent} />
-          1차
-        </Text>
-      </Table.Td>
-      <Table.Td>
-        <Text style={awardTextStyle} typo="body2">
-          <AwardIcon disabled={!isSecondRoundOutstandingStudent} />
-          2차
-        </Text>
-      </Table.Td>
-      <Table.Td>{name}</Table.Td>
-      <Table.Td>{studentId}</Table.Td>
-      <Table.Td>{discordUsername}</Table.Td>
-      <Table.Td>{nickname}</Table.Td>
-      <Table.Td>
-        <TextButton
-          asProp={Link}
-          href={githubLink || ""}
-          style={textButtonStyle}
-          text={githubLink}
-        />
-      </Table.Td>
-      <StudyTasksTds tasks={studyTasks} />
-      <Table.Td>{formatNumberToPercent(assignmentRate)}</Table.Td>
-      <Table.Td>{formatNumberToPercent(attendanceRate)}</Table.Td>
-    </Table.Tr>
-  );
-};
-
-const StudyTasksThs = ({
-  tasks,
-}: {
-  tasks: (
-    | StudyTaskResponseDto<"ASSIGNMENT">
-    | StudyTaskResponseDto<"ATTENDANCE">
-  )[];
-}) => {
-  return tasks.map((task) => {
-    const { week, taskType } = task;
-    return (
-      <Table.Th key={taskType + week}>
-        {taskType === "ATTENDANCE" ? `${week}주차 출석` : `${week}주차 과제`}
-      </Table.Th>
-    );
-  });
-};
-
-const StudyTasksTds = ({
-  tasks,
-}: {
-  tasks: (
-    | StudyTaskResponseDto<"ASSIGNMENT">
-    | StudyTaskResponseDto<"ATTENDANCE">
-  )[];
-}) => {
-  return tasks.map((task) => {
-    const { week, taskType } = task;
-    return (
-      <Table.Td key={taskType + week}>
-        <TaskTag task={task} />
-      </Table.Td>
-    );
-  });
-};
-
-const textButtonStyle: CSSProperties = {
-  width: "fit-content",
-  padding: 0,
-};
-
-const awardTextStyle: CSSProperties = {
-  display: "flex",
-  gap: "0.25rem",
-  alignItems: "center",
-};
 export default StudentList;
